✨ feat: 아이템 상세보기 이벤트리스너 추가
